Update Calendar.Js
Updated how the calendar displays events. Uses coloured circle which are currently randomly generated each time they are read.
Displaying content as propsed by @iman-hussain, changes to positioning, font, etc will be made in the future once discussed with the team.
Future event coloring could be set in the database or could be set and received from the users calendar.