State management and peristent storage
Added state management in the form of react-redux. This will make accessing and saving reusable information more efficient. This will remove the need for constantly calling the google calendar API.
Added redux-persist so as to store the users information in local storage, which will then be retrieve and used by react-redux, so that the user do not need to log back in again.