Enable user login (currently only matches email and none empty password)
Enabled user login whereby the users email is sent to the backend and a true or false status will be recieved alongside the users details.

The users details are then stored using react-redux toolkit, which can then be used throughout the application.

Upon updating the database with a password field this can be easily changed to use verify the users password alongside their email.