Moving toward DB communication over Google API
Changed the way events are retrieved from using the Google API to connecting to the database.

Current issues involve the need for database field updates, mainly event color as a visual aid.

Removed tick fuctions where the page was reloading every 5 seconds to update the Google calendar API information, it now only needs to load when there is a change.