Created a sign up page
Create a sign up page which sends the users information to the backend, which then creates a user account.

Testing and a password field is needed to enhance this page and ensure it's working as intended and ensure it can handle errors when connecting to the API or recieving a bad request.