Fixed empty booking error and duplicate bookings
Fixed an issue where an empty booking array could crash the application. For this i just check if the array exists and if their is any elements within the array.

Fixed an issue where duplicate events would be shown on the calendar. For this i removed the ability to fetch events from the google calendar and inplace i am using the booking information from the database.

closes #24